Edit Action in Customers Controller
Edit takes in an id that is put within the customer variable that accesses the Customers table in the database context. If there is no customer, it returns 404 Not found. Else, it will generate a CustomerFormViewModel that takes in the customer variable declared at the beginning and assigns it to Customer and then looks in the context for Membership Types and converts it to a list. 

Then it returns the view "Customer form" using the viewModel declared previously.